// web/src/api/generated/model/execution.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Firetower
 * The Firetower control plane: API, scheduling, and worker transports.
 * OpenAPI spec version: 0.30.1
 */

export type Execution = typeof Execution[keyof typeof Execution];


export const Execution = {
  container: 'container',
  host: 'host',
} as const;